moved folder appUser to the correct package
